Move into package structure; add pyproject.toml
