+++ .dockerignore
@@ -2,3 +2,11 @@ example/
 __pycache__
 **.mzml
 test/
+
+# Environment
+.venv
+
+# Builds
+build/
+dist/
+*.egg-info/
